% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Functions.R
\name{aggData}
\alias{aggData}
\title{Aggregate data with optional covariates}
\usage{
aggData(
  data.total = NULL,
  covariates = NULL,
  group_by_cols = NULL,
  .cat = FALSE,
  timing = FALSE,
  ...
)
}
\description{
Aggregate data with optional covariates
}
