% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_data_function_revision.R
\name{api_data}
\alias{api_data}
\title{Title}
\usage{
api_data(
  periods = NA,
  periodType = "Monthly",
  YrsPrevious = 1,
  orgUnits = "LEVEL-1",
  elements = NA,
  baseurl = NA,
  username = NULL,
  password = NULL,
  formula = NA,
  parallel = FALSE,
  print = TRUE,
  update = FALSE,
  check_previous_years = 2,
  previous_dataset_file = "",
  prev.data = NA,
  level1.id = NA,
  dir = country.dir,
  shinyApp = FALSE,
  childOnly = FALSE,
  ...
)
}
\arguments{
\item{periods}{specify period}

\item{periodType}{Monthly or Weekly}

\item{YrsPrevious}{Number of years of data}

\item{orgUnits}{Level of data (default = LEVEL-1)}

\item{elements}{data elements}

\item{baseurl}{DHIS2 instance url}

\item{username}{user name}

\item{password}{password}

\item{formula}{formula with data elements}

\item{parallel}{make prallel requests}

\item{print}{print messages}

\item{update}{update}

\item{check_previous_years}{if previous_dataset_file, number of years to check}

\item{prev.data}{.}

\item{level1.id}{when comparing current data with previous}

\item{dir}{directory where downloads saved}

\item{shinyApp}{run in shine (FALSE)}

\item{childOnly}{Data only from child facilities}

\item{...}{}
}
\description{
Title
}
