% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/climate_functions.R
\name{chirps_country_options}
\alias{chirps_country_options}
\title{African country ISO-3 codes for GADM / CHIRPS analysis}
\usage{
chirps_country_options()
}
\value{
Named character vector mapping country name to ISO-3 code.
}
\description{
African country ISO-3 codes for GADM / CHIRPS analysis
}
