% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Functions.R
\name{cleanedData}
\alias{cleanedData}
\title{Apply cleaning and filtering to a dataset}
\usage{
cleanedData(
  d,
  .effectiveLeaf = TRUE,
  source = "Original",
  error = NULL,
  algorithm = "seasonal3",
  .cat = FALSE
)
}
\arguments{
\item{d}{data frame; the dataset to clean}

\item{.effectiveLeaf}{logical; filter to effective leaf org units}

\item{source}{character; data source label}

\item{...}{additional arguments}
}
\value{
cleaned data frame
}
\description{
Apply cleaning and filtering to a dataset
}
