% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Functions.R
\name{dataPeriod}
\alias{dataPeriod}
\title{Determine the period type used in a dataset}
\usage{
dataPeriod(data1, .cat = FALSE)
}
\arguments{
\item{data1}{data frame; dataset with a time column}

\item{.cat}{logical; if TRUE print progress messages}
}
\value{
character; "Monthly" or "Weekly"
}
\description{
Determine the period type used in a dataset
}
