% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Functions.R
\name{dataTotal}
\alias{dataTotal}
\title{Merge and aggregate datasets across org units and periods}
\usage{
dataTotal(
  data = NULL,
  group_by_cols = NULL,
  period = "Month",
  startMonth = NULL,
  endMonth = NULL,
  dataSets = NULL,
  sum.data = TRUE,
  mean.merge = FALSE,
  covariates = NULL,
  .cat = TRUE
)
}
\description{
Merge and aggregate datasets across org units and periods
}
