% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqa_functions.R
\name{dqa_consistency_plot}
\alias{dqa_consistency_plot}
\title{Annual Consistency Chart}
\usage{
dqa_consistency_plot(consistency_data, text_size = 18)
}
\arguments{
\item{consistency_data}{Output of \code{\link[=dqa_consistency]{dqa_consistency()}}.}

\item{text_size}{Numeric. Base text size (default: \code{18}).}
}
\value{
A ggplot object.
}
\description{
Annual Consistency Chart
}
