% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqa_functions.R
\name{dqa_consistency_table}
\alias{dqa_consistency_table}
\title{Per-Rule Summary Table}
\usage{
dqa_consistency_table(consistency_data)
}
\arguments{
\item{consistency_data}{Output of \code{\link[=dqa_consistency]{dqa_consistency()}}.}
}
\value{
A tibble with one row per rule.
}
\description{
Per-Rule Summary Table
}
