% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqa_functions.R
\name{dqa_facility_status_plot}
\alias{dqa_facility_status_plot}
\title{Stacked bar chart of Active vs Inactive facilities by region}
\usage{
dqa_facility_status_plot(status_data, text_size = 16)
}
\arguments{
\item{status_data}{Output of \code{\link[=dqa_facility_status]{dqa_facility_status()}}.}

\item{text_size}{Numeric base text size (default 16).}
}
\value{
A ggplot object.
}
\description{
Stacked bar chart of Active vs Inactive facilities by region
}
