% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqa_functions.R
\name{dqa_facility_status_table}
\alias{dqa_facility_status_table}
\title{Summary table: region | total | active | inactive | \% active}
\usage{
dqa_facility_status_table(status_data)
}
\arguments{
\item{status_data}{Output of \code{\link[=dqa_facility_status]{dqa_facility_status()}}.}
}
\value{
A data.frame suitable for \code{DT::datatable()}.
}
\description{
Summary table: region | total | active | inactive | \% active
}
