% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqa_functions.R
\name{dqa_mase_plot}
\alias{dqa_mase_plot}
\title{Plot MASE Summary}
\usage{
dqa_mase_plot(data)
}
\arguments{
\item{data}{Output of \code{\link[=dqa_mase]{dqa_mase()}}.}
}
\value{
A ggplot object.
}
\description{
Plot MASE Summary
}
