% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqa_functions.R
\name{dqa_outliers}
\alias{dqa_outliers}
\title{Summarise Yearly Outlier Flags from DQA Data}
\usage{
dqa_outliers(yearly.outlier.summary)
}
\arguments{
\item{yearly.outlier.summary}{Output of \code{yearly.outlier.summary()}.}
}
\value{
A tibble with percent-no-error columns added.
}
\description{
Summarise Yearly Outlier Flags from DQA Data
}
