% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqa_functions.R
\name{dqa_reporting}
\alias{dqa_reporting}
\title{Count Consistently Reporting Facilities per Year}
\usage{
dqa_reporting(
  dqa_data,
  missing_reports = 0,
  count.any = TRUE,
  .cat = FALSE,
  ...
)
}
\arguments{
\item{dqa_data}{tsibble of prepared data.}

\item{missing_reports}{Integer. Allowed missing months (default: \code{0}).}

\item{count.any}{Logical. Count facilities reporting any data element
(default: \code{TRUE}).}

\item{.cat}{Logical. Print progress messages (default: \code{FALSE}).}
}
\value{
Numeric vector with count of consistently reporting facilities per year.
}
\description{
Count Consistently Reporting Facilities per Year
}
