% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqa_functions.R
\name{dqa_reporting_plot}
\alias{dqa_reporting_plot}
\title{Plot Percent of Facilities Consistently Reporting per Year}
\usage{
dqa_reporting_plot(data, text_size = 18)
}
\arguments{
\item{data}{Output of \code{\link[=dqaPercentReporting]{dqaPercentReporting()}}.}

\item{text_size}{Numeric. Base text size for the plot (default: \code{18}).}
}
\value{
A ggplot object.
}
\description{
Plot Percent of Facilities Consistently Reporting per Year
}
