% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqa_functions.R
\name{dqa_swape_plot}
\alias{dqa_swape_plot}
\alias{dqa_mase_plot}
\title{Plot SWAPE Summary (Minimum Detectable Change)}
\usage{
dqa_swape_plot(data)

dqa_mase_plot(data)
}
\arguments{
\item{data}{Output of \code{\link[=dqa_swape]{dqa_swape()}}.}
}
\value{
A ggplot object.
}
\description{
Plot SWAPE Summary (Minimum Detectable Change)
}
