% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/api_data_function_revision.R
\name{find_lowest_nonnull}
\alias{find_lowest_nonnull}
\title{Find the lowest non-null level value from a data request selection}
\usage{
find_lowest_nonnull(selected)
}
\arguments{
\item{selected}{list; named list with elements \code{level2} through \code{level5} (NULL when not selected)}
}
\description{
Find the lowest non-null level value from a data request selection
}
