% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Functions.R
\name{getForecast}
\alias{getForecast}
\title{Extract forecast values for a specific model and replicate}
\usage{
getForecast(
  forecastData,
  model = NULL,
  model.string = NULL,
  transform = TRUE,
  lambda = 0.5,
  covariates = NULL,
  horizon = 12,
  bootstrap = FALSE,
  Reps = 1000,
  future.seed = TRUE,
  split = "None",
  agg_level = NULL,
  agg_method = "None",
  .period = "Month",
  eval_date = as.yearmonth("Jan 2021"),
  .cat = FALSE
)
}
\description{
Extract forecast values for a specific model and replicate
}
