% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Functions.R
\name{getLevelNames}
\alias{getLevelNames}
\title{Get level names from org units}
\usage{
getLevelNames(orgUnits, .cat = FALSE)
}
\arguments{
\item{orgUnits}{data frame of org units with level and levelName columns}

\item{.cat}{logical; if TRUE print progress messages}
}
\value{
character vector of level names in ascending order
}
\description{
Get level names from org units
}
