% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Functions.R
\name{htsData}
\alias{htsData}
\title{Aggregate data using hierarchical time-series structure}
\usage{
htsData(
  data = NULL,
  hts = TRUE,
  hts_formula = NULL,
  covariates = "",
  group_by_cols = NULL,
  .cat = FALSE,
  timing = FALSE,
  ...
)
}
\description{
Aggregate data using hierarchical time-series structure
}
