% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Functions.R
\name{htsFormula}
\alias{htsFormula}
\title{Build formula for hierarchical time-series aggregation}
\usage{
htsFormula(
  hts = TRUE,
  levelNames = NULL,
  agg_level = NULL,
  all.levels = FALSE,
  num_facilities = NULL,
  num_datasets = NULL,
  split = "None",
  .cat = FALSE
)
}
\arguments{
\item{hts}{logical; use hierarchical time-series aggregation}

\item{levelNames}{character vector; administrative level names}

\item{agg_level}{character; target aggregation level name}

\item{all.levels}{logical; include all levels in the formula}

\item{num_facilities}{integer; number of facilities in the dataset}

\item{num_datasets}{integer; number of datasets in the dataset}

\item{split}{character; split variable ("None" or a column name)}

\item{.cat}{logical; print progress messages to the console}
}
\description{
Build formula for hierarchical time-series aggregation
}
