% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/combine_functions.R
\name{load_combine_definition}
\alias{load_combine_definition}
\title{Load a combination definition from a companion RDS file}
\usage{
load_combine_definition(file_path)
}
\arguments{
\item{file_path}{Full path to a \code{Combinations_*.rds} file.}
}
\value{
Definition list.
}
\description{
Load a combination definition from a companion RDS file
}
