% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Functions.R
\name{MAPE}
\alias{MAPE}
\title{Compute mean absolute percentage error}
\usage{
MAPE(preForecast, trend.data, period = "Month", var = ".mean", .cat = FALSE)
}
\description{
Compute mean absolute percentage error
}
