% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/outlier_summary.R
\name{outlier.summary.chart}
\alias{outlier.summary.chart}
\title{Plot Monthly Outlier Summaries by Year and Algorithm}
\usage{
outlier.summary.chart(grand.data.error.totals)
}
\arguments{
\item{grand.data.error.totals}{Output of \code{\link[=monthly.outlier.summary]{monthly.outlier.summary()}}.}
}
\value{
A ggplot object.
}
\description{
Plot Monthly Outlier Summaries by Year and Algorithm
}
