% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Functions.R
\name{plotTrends}
\alias{plotTrends}
\title{Plot time-series trends with forecast overlay}
\usage{
plotTrends(
  trend.data,
  scale = FALSE,
  legend = FALSE,
  label = FALSE,
  facet_split = FALSE,
  facet_vars = NULL,
  facet_admin = FALSE,
  agg_level = "National",
  pre_evaluation = FALSE,
  evaluation = FALSE,
  horizon = 12,
  eval_date = as.yearmonth("Jan 2021", "\%Y \%b"),
  pe = TRUE,
  .cat = FALSE,
  ...
)
}
\description{
Plot time-series trends with forecast overlay
}
