% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Functions.R
\name{pre_impact_fit}
\alias{pre_impact_fit}
\title{Fit pre-intervention impact models}
\usage{
pre_impact_fit(
  ml.data = ml.data,
  startingMonth = "Jan 2015",
  endingMonth = NULL,
  selected.only = FALSE,
  missing_reports = 0,
  split = NULL,
  evaluation_month = "Dec 2018",
  error = "seasonal3",
  hts = TRUE,
  aggregate = TRUE,
  levelNames = NULL,
  agg_level = NULL,
  pi_levels = 89,
  model = "ARIMA",
  modelSpecs = NULL,
  transform = FALSE,
  covariates = NULL,
  simulate = FALSE,
  bootstrap = FALSE,
  times = 100,
  remove.aggregate = TRUE,
  .cat = TRUE
)
}
\description{
Fit pre-intervention impact models
}
