% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Functions.R
\name{selectedData}
\alias{selectedData}
\title{Filter and select data by reporting level and categories}
\usage{
selectedData(
  data1,
  levelNames = NULL,
  data_categories = NULL,
  alwaysReporting = TRUE,
  missing_reports = 0,
  reportingSelectedOUs = NULL,
  startingMonth = NULL,
  endingMonth = NULL,
  level = "leaf",
  level2 = NULL,
  level3 = NULL,
  level4 = NULL,
  level5 = NULL,
  .cat = FALSE,
  ...
)
}
\description{
Filter and select data by reporting level and categories
}
