% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Functions.R
\name{tsModel}
\alias{tsModel}
\title{Fit time-series models across replicates}
\usage{
tsModel(
  trend.data,
  evaluation_month,
  model = "ARIMA",
  formula.string = NULL,
  period = "Month",
  .cat = TRUE,
  ...
)
}
\description{
Fit time-series models across replicates
}
