% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{Week_Year}
\alias{Week_Year}
\title{Convert a DHIS2 Weekly Period Code to yearweek}
\usage{
Week_Year(x)
}
\arguments{
\item{x}{Character vector of DHIS2 weekly period codes.}
}
\value{
A \code{tsibble::yearweek} vector.
}
\description{
Parses DHIS2 weekly period codes to \code{\link[tsibble:year-week]{tsibble::yearweek()}} objects.
}
\examples{
Week_Year("2019W01")
}
