% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqa_functions.R
\name{yearly.outlier.summary_plot}
\alias{yearly.outlier.summary_plot}
\title{Plot Yearly Outlier Summary}
\usage{
yearly.outlier.summary_plot(data, text_size = 18, label_size = 6)
}
\arguments{
\item{data}{Output of \code{\link[=dqa_outliers]{dqa_outliers()}}.}

\item{text_size}{Numeric. Base text size (default: \code{18}).}

\item{label_size}{Numeric. Geom text size (default: \code{6}).}
}
\value{
A ggplot object.
}
\description{
Plot Yearly Outlier Summary
}
