% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Functions.R
\name{as.yearmonth}
\alias{as.yearmonth}
\title{functions for data manipulation
as.yearmonth: convert character string to yearmonth object}
\usage{
as.yearmonth(date.string, fmt = "\%B\%Y")
}
\arguments{
\item{date.string}{character string representing a date}

\item{fmt}{format of the date string, default is "\%B\%Y" (e.g. "January2020")}
}
\value{
a yearmonth object
}
\description{
functions for data manipulation
as.yearmonth: convert character string to yearmonth object
}
\examples{
as.yearmonth("Jan2020")
}
