% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/dqa_functions.R
\name{dqaPercentReporting}
\alias{dqaPercentReporting}
\title{Compute Percent of Facilities Consistently Reporting per Year}
\usage{
dqaPercentReporting(dqa_data, .cat = FALSE)
}
\arguments{
\item{dqa_data}{tsibble of prepared data.}

\item{.cat}{Logical. Print progress messages (default: \code{FALSE}).}
}
\value{
A tibble with columns \code{Year}, \code{n_frequently_reporting},
\code{n_facilities}, \code{pr}, \code{label}.
}
\description{
Compute Percent of Facilities Consistently Reporting per Year
}
