% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Functions.R
\name{groupByCols}
\alias{groupByCols}
\title{Build group-by column vector for data aggregation}
\usage{
groupByCols(
  selected = TRUE,
  dataset = TRUE,
  orgUnit = TRUE,
  data = FALSE,
  period = NULL,
  hts = FALSE,
  agg_level = NULL,
  levelNames = NULL,
  split = NULL,
  .cat = FALSE
)
}
\description{
Build group-by column vector for data aggregation
}
