% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Functions.R
\name{htsFormula}
\alias{htsFormula}
\title{Build formula for hierarchical time-series aggregation}
\usage{
htsFormula(
  hts = TRUE,
  levelNames = NULL,
  agg_level = NULL,
  all.levels = FALSE,
  num_facilities = NULL,
  num_datasets = NULL,
  split = "None",
  .cat = FALSE
)
}
\description{
Build formula for hierarchical time-series aggregation
}
