% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Functions.R
\name{key.mape}
\alias{key.mape}
\title{Compute key MAPE values for forecast accuracy summary}
\usage{
key.mape(
  preForecast,
  trend.data,
  period = "Month",
  split = "None",
  agg_level = NULL,
  horizon = 12,
  .cat = FALSE
)
}
\description{
Compute key MAPE values for forecast accuracy summary
}
