% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Functions.R
\name{key.mpe}
\alias{key.mpe}
\title{Compute key mean percentage error values for forecast evaluation}
\usage{
key.mpe(
  Forecast_data,
  test_data,
  period = "Month",
  .split = "None",
  by_month = FALSE,
  agg_level = NULL,
  horizon = NULL,
  var = ".mean",
  .cat = FALSE
)
}
\description{
Compute key mean percentage error values for forecast evaluation
}
