% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{mg2_data_ext}
\alias{mg2_data_ext}
\title{Return the Preferred Data File Extension}
\usage{
mg2_data_ext()
}
\value{
\code{"fst"} or \code{"rds"}.
}
\description{
Returns \code{"fst"} when the \code{fst} package is installed (faster I/O), otherwise
\code{"rds"}. Use this to build filenames for \code{\link[=save_file]{save_file()}}.
}
