% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Functions.R
\name{model_formula}
\alias{model_formula}
\title{Build model formula string for time-series fitting}
\usage{
model_formula(
  model = "ARIMA",
  modelSpecs = NULL,
  transform = FALSE,
  period = "Month",
  covariates = NULL,
  .cat = FALSE
)
}
\description{
Build model formula string for time-series fitting
}
