% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Functions.R
\name{mostFrequentReportingOUs}
\alias{mostFrequentReportingOUs}
\title{Find most frequently reporting org units}
\usage{
mostFrequentReportingOUs(
  data,
  endingMonth = NULL,
  startingMonth = NULL,
  period = NULL,
  missing_reports = 0,
  count.any = TRUE,
  data_categories = NULL,
  testing = FALSE,
  .cat = FALSE
)
}
\description{
Find most frequently reporting org units
}
