% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Functions.R
\name{trendData}
\alias{trendData}
\title{Prepare trend data for modelling and forecasting}
\usage{
trendData(
  .d = data.hts,
  reportingSelectedOUs = NULL,
  period = "Month",
  startingMonth = NULL,
  endingMonth = NULL,
  selected.only = TRUE,
  num_facilities = NULL,
  num_datasets = 1,
  levelNames = NULL,
  agg_level = NULL,
  split = "None",
  covariates = NULL,
  remove.aggregate = TRUE,
  scale = FALSE,
  testing = FALSE,
  .cat = FALSE
)
}
\description{
Prepare trend data for modelling and forecasting
}
