% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Functions.R
\name{tsForecast}
\alias{tsForecast}
\title{Generate time-series forecasts from fitted models}
\usage{
tsForecast(
  trend.data,
  Model,
  horizon = 12,
  evaluation_month = NULL,
  period = "Month",
  covariates = NULL,
  split = NULL,
  agg_level = NULL,
  simulate = FALSE,
  bootstrap = FALSE,
  times = 100,
  pi_levels = 0.89,
  .cat = FALSE
)
}
\description{
Generate time-series forecasts from fitted models
}
