% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Functions.R
\name{tsPreForecast}
\alias{tsPreForecast}
\title{Generate pre-intervention forecasts from fitted models}
\usage{
tsPreForecast(
  trend.data,
  preModel,
  horizon = 12,
  evaluation_month = NULL,
  period = "Month",
  covariates = NULL,
  split = NULL,
  agg_level = NULL,
  prob = FALSE,
  pi_levels = 0.89,
  .cat = FALSE
)
}
\description{
Generate pre-intervention forecasts from fitted models
}
