% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data_Functions.R
\name{tsPreModel}
\alias{tsPreModel}
\title{Fit pre-intervention time-series models}
\usage{
tsPreModel(
  trend.data,
  evaluation_month,
  model = "ARIMA",
  formula.string = NULL,
  period = "Month",
  .cat = TRUE,
  ...
)
}
\description{
Fit pre-intervention time-series models
}
